Updated AStar and LazySP
Allow AstarLogging to take either a const Graph, or a non-const Graph (i.e. Astar can change the graph, presumably just the edge validity).

Change LazySP to continue valididating the path if evaluating an edge did not increase the cost. The edge evaluations should be equivalent, but fewer iteration of A* are needed